Renaming and cleaning
