+++ fAIth/scripts/embeddings_test.py
@@ -10,7 +10,7 @@ PROJECT_ROOT = Path(__file__).resolve().parents[1]
 if str(PROJECT_ROOT) not in sys.path:
     sys.path.insert(0, str(PROJECT_ROOT))
 
-from ai.embedding.embedding import Embedding
+from ai.vdb.embedding import Embedding
 load_dotenv()
 
 model_id = os.getenv("EMBEDDING_MODEL_ID", "")
